+++ markup/css/main.css
@@ -1409,3 +1409,14 @@ main {
   top: 50%;
   -webkit-transform: translateY(-50%);
           transform: translateY(-50%); }
+
+
+.small-film-card video {
+  position: absolute;
+  vertical-align: top;
+    width: 100%;
+    height: 100%;
+    -o-object-fit: cover;
+       font-family: "object-fit:cover";
+       object-fit: cover;
+}
